Set basic verification on some methods

+++ src/main/java/be/ordina/blog/controller/PostRestController.java
@@ -4,10 +4,12 @@ import be.ordina.blog.model.Comment;
 import be.ordina.blog.model.Post;
 import be.ordina.blog.service.PostService;
 import io.swagger.annotations.ApiOperation;
+import org.apache.http.HttpException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 
 import static org.springframework.web.bind.annotation.RequestMethod.*;
@@ -45,7 +47,7 @@ public class PostRestController {
             value = "Update a post",
             notes = "id should be provided in the request body of the post")
     @RequestMapping(method = PUT)
-    public void updatePost(@PathVariable long postId, @RequestBody Post post) {
+    public void updatePost(@PathVariable long postId, @RequestBody Post post , Principal principal) {
         logger.info("Request to update post", postId);
         postService.updatePost(post);
     }
@@ -54,7 +56,7 @@ public class PostRestController {
             value = "Update a post",
             notes = "postId from the URL is used to select the post")
     @RequestMapping(value = "/{postId}", method = PUT)
-    public void updatePostById(@PathVariable long postId, @RequestBody Post post) {
+    public void updatePostById(@PathVariable long postId, @RequestBody Post post , Principal principal) {
         logger.info("Request to update post with id {}", postId);
         post.setId(postId);
         postService.updatePost(post);
@@ -64,36 +66,37 @@ public class PostRestController {
             value = "Deletes a post",
             notes = "This method is far less performing than /topics/{topicId}/posts/{postsId}. Use with caution.")
     @RequestMapping(value = "/{postId}", method = DELETE)
-    public void deletePostById(@PathVariable long postId) {
+    public void deletePostById(@PathVariable long postId  , Principal principal) {
         postService.deletePostById(postId);
     }
 
     @ApiOperation(
             value = "Get all comment of a post")
     @RequestMapping(value = "/{postId}/comments", method = GET)
-    public List<Comment> getCommentsForPost(@PathVariable long postId) {
+    public List<Comment> getCommentsForPost(@PathVariable long postId  , Principal principal) {
         return postService.getCommentsForPost(postId);
     }
 
     @ApiOperation(
             value = "Get a comment of a post")
     @RequestMapping(value = "/{postId}/comments/{commentId}", method = GET)
-    public Comment getCommentForPost(@PathVariable long postId, @PathVariable long commentId) {
+    public Comment getCommentForPost(@PathVariable long postId, @PathVariable long commentId  , Principal principal) {
         return postService.getCommentForPost(commentId, postId);
     }
 
     @ApiOperation(
             value = "Create a new comment for a post")
     @RequestMapping(value = "/{postId}/comments", method = POST)
-    public void addCommentToPost(@PathVariable long postId, @RequestBody Comment comment) {
+    public void addCommentToPost(@PathVariable long postId, @RequestBody Comment comment, Principal principal) {
         postService.addCommentToPost(comment, postId);
     }
 
+    // moderator
     @ApiOperation(
             value = "Update a comment from a post",
             notes = "id should be provided in the request body comment")
     @RequestMapping(value = "/{postId}/comments", method = PUT)
-    public void updateCommentForPost(@PathVariable long postId, @RequestBody Comment comment) {
+    public void updateCommentForPost(@PathVariable long postId, @RequestBody Comment comment, Principal principal) throws ForbiddenException {
         postService.updateCommentForPost(comment, postId);
     }
 
@@ -101,7 +104,7 @@ public class PostRestController {
             value = "Update a comment from a post",
             notes = "commentId in URL will be used for selecting the comment")
     @RequestMapping(value = "/{postId}/comments/{commentId}", method = PUT)
-    public void updateCommentForPost(@PathVariable long postId, @PathVariable long commentId, @RequestBody Comment comment) {
+    public void updateCommentForPost(@PathVariable long postId, @PathVariable long commentId, @RequestBody Comment comment, Principal principal) {
         comment.setId(commentId);
         postService.updateCommentForPost(comment, postId);
     }
@@ -109,7 +112,10 @@ public class PostRestController {
     @ApiOperation(
             value = "Delete a comment from a post")
     @RequestMapping(value = "/{postId}/comments/{commentId}", method = DELETE)
-    public void deleteCommentForPost(@PathVariable long postId, @PathVariable long commentId) {
+    public void deleteCommentForPost(@PathVariable long postId, @PathVariable long commentId, Principal principal)throws ForbiddenException {
+        if (!principal.toString().contains("ROLE_moderator")){
+            throw new ForbiddenException("This user doesn't have moderator rights!");
+        }
         postService.deleteCommentForPost(commentId, postId);
     }
 }
