+++ src/main/java/be/ordina/blog/controller/TopicRestController.java
@@ -7,6 +7,7 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.models.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.nio.file.AccessDeniedException;
 import java.security.Principal;
 import java.util.List;
 
@@ -40,8 +41,7 @@ public class TopicRestController {
     @ApiOperation(
             value = "Create a new topic")
     @RequestMapping(method = POST)
-    public void addTopic(@RequestBody Topic topic, Principal principal , Model model) {
-        System.out.println(principal.toString());
+    public void addTopic(@RequestBody Topic topic, Principal principal) {
         topicService.addTopic(topic);
     }
 
@@ -49,7 +49,7 @@ public class TopicRestController {
             value = "update a topic",
             notes = "id is expected in the request body topic")
     @RequestMapping(method = PUT)
-    public void updateTopic(@RequestBody Topic topic) {
+    public void updateTopic(@RequestBody Topic topic,Principal principal) {
         topicService.updateTopic(topic);
     }
 
@@ -57,7 +57,7 @@ public class TopicRestController {
             value = "Update a topic",
             notes = "topicId in the URL is used to select the topic")
     @RequestMapping(value="/{topicId}" , method = PUT)
-    public void updateTopic(@PathVariable long topicId, @RequestBody Topic topic) {
+    public void updateTopic(@PathVariable long topicId, @RequestBody Topic topic , Principal principal) {
         topic.setId(topicId);
         topicService.updateTopic(topic);
     }
@@ -65,28 +65,31 @@ public class TopicRestController {
     @ApiOperation(
             value = "Delete a topic")
     @RequestMapping(value="/{topicId}" ,method = DELETE)
-    public void deleteTopicById(@PathVariable Long topicId) {
+    public void deleteTopicById(@PathVariable Long topicId , Principal principal) throws ForbiddenException {
+        if (!principal.toString().contains("ROLE_admin")){
+            throw new ForbiddenException("This user doesn't have admin rights!");
+        }
         topicService.deleteTopicById(topicId);
     }
 
     @ApiOperation(
             value = "Get all post of a topic")
     @RequestMapping(value="/{topicId}/posts", method = GET)
-    public List<Post> getPostsForTopic(@PathVariable long topicId) {
+    public List<Post> getPostsForTopic(@PathVariable long topicId ,Principal principal) {
         return topicService.getPostsForTopic(topicId);
     }
 
     @ApiOperation(
             value = "Get a post of a topic")
     @RequestMapping(value="/{topicId}/posts/{postId}", method = GET)
-    public Post getPostForTopic(@PathVariable long topicId, @PathVariable long postId) {
+    public Post getPostForTopic(@PathVariable long topicId, @PathVariable long postId,Principal principal) {
         return topicService.getPostForTopic(postId, topicId);
     }
 
     @ApiOperation(
             value = "Create a new post for a topic")
     @RequestMapping(value="/{topicId}/posts", method = POST)
-    public void addPostForTopic(@PathVariable long topicId, @RequestBody Post post) {
+    public void addPostForTopic(@PathVariable long topicId, @RequestBody Post post, Principal principal) {
         topicService.addPostForTopic(post, topicId);
     }
 
@@ -94,7 +97,7 @@ public class TopicRestController {
             value = "Update a post from a topic",
             notes = "id is expected in the request body post")
     @RequestMapping(value="/{topicId}/posts", method = PUT)
-    public void updatePostForTopic(@PathVariable long topicId, @RequestBody Post post) {
+    public void updatePostForTopic(@PathVariable long topicId, @RequestBody Post post , Principal principal) {
         topicService.updatePostForTopic(post, topicId);
     }
 
@@ -102,7 +105,7 @@ public class TopicRestController {
             value = "Update a post from a topic",
             notes = "postId from URL is used to select the post")
     @RequestMapping(value="/{topicId}/posts/{postId}", method = PUT)
-    public void updatePostForTopic(@PathVariable long topicId, @PathVariable long postId, @RequestBody Post post) {
+    public void updatePostForTopic(@PathVariable long topicId, @PathVariable long postId, @RequestBody Post post , Principal principal) {
         post.setId(postId);
         topicService.updatePostForTopic(post, topicId);
     }
@@ -110,7 +113,7 @@ public class TopicRestController {
     @ApiOperation(
             value = "Delete a post from a topic")
     @RequestMapping(value = "/{topicId}/posts/{postId}", method = DELETE)
-    public void deletePostForTopic(@PathVariable long topicId, @PathVariable long postId) {
+    public void deletePostForTopic(@PathVariable long topicId, @PathVariable long postId  , Principal principal) {
         topicService.removePostForTopic(postId, topicId);
     }
 }
